Workaround alignment bug in Grub 2 that prevented boot. Code cleanup. Upgrade to Rust master. Other fixes.

+++ arch/x86_64/idt_handlers.rs
@@ -1,11 +1,9 @@
 use super::IDT_SIZE;
 
-extern
-{
+extern {
     static _asm_irq_handler_array: [u64, ..IDT_SIZE];
 }
 
-pub fn get_irq_handler(num: u16) -> u64
-{
-    _asm_irq_handler_array[num]
+pub fn get_irq_handler(num: u16) -> u64 {
+    _asm_irq_handler_array[num as uint]
 }
